backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
